pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stock-ai"
version = "0.1.0"
description = "Personal stock classification AI project for Japanese equities."
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "ryo" }
]
dependencies = [
  "joblib>=1.4",
  "lightgbm>=4.5",
  "pandas>=2.2",
  "PyYAML>=6.0.1",
  "requests>=2.31",
  "scikit-learn>=1.5",
  "yfinance>=0.2.54",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.6",
]

[project.scripts]
stock-ai = "stock_ai.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
